# openapi.yaml
openapi: 3.1.0
info:
  title: DiArMaqAr API
  version: 1.0.0
  description: |
    Digital Arabic Maqām Archive (DiArMaqAr) — programmatic access to Arabic maqāmāt, ajnās, and tuning systems data

    Key features:
    - Comprehensive maqāmāt and ajnās database with bibliographic source attribution
    - Tuning systems spanning 1000+ years with multiple starting note contexts
    - Comprehensive pitch class data formats (fractions ratios, decimal ratio, string length, fret divisions, frequency, cents, MIDI notes, and more)
    - Tuning system sensitive transpositions and modulations via al-Shawwā's 1946 modulation guidelines

  contact:
    name: DiArMaqAr 
    url: https://github.com/Music-Intelligence-Lab/DiArMaqAr
  license:
    name: CC-BY-NC-SA 4.0
    url: https://creativecommons.org/licenses/by-nc-sa/4.0/
servers:
  - url: https://diarmaqar.netlify.app/api
    description: Production API
  - url: http://localhost:3000/api
    description: Local development
paths:
...
  /tuning-systems/{id}/{startingNote}/pitch-classes:
    get:
      tags:
        - Tuning Systems
      summary: Get Tuning System Details
      description: |
        Get Tuning System Details - Returns all pitch classes for a specific tuning system and starting note.
        
        This endpoint is essential for tuning system operations, providing comprehensive pitch class data
        across all octaves with full formatting options.
      operationId: getTuningSystemPitchClasses
      parameters:
        - name: id
          in: path
          required: true
          description: Tuning system identifier
          schema:
            type: string
          example: "IbnSina-(1037)"
        - name: startingNote
          in: path
          required: true
          description: Tuning system starting note name (URL-safe, diacritics-insensitive)
          schema:
            type: string
          example: "yegah"
        - name: pitchClassDataType
          in: query
          required: false
          description: Pitch class data format (optional, defaults to "all")
          schema:
            type: string
            enum:
              - all
              - englishName
              - fraction
              - cents
              - decimalRatio
              - stringLength
              - frequency
              - abjadName
              - fretDivision
              - midiNoteNumber
              - midiNoteDeviation
              - centsDeviation
              - referenceNoteName
          default: "all"
          example: "cents"
        - name: octave
          in: query
          required: false
          description: |
            Filter by octave number. Use "all" to return pitch classes from all octaves (default).
            Use a specific octave number (0, 1, 2, 3) to filter to that octave only.
          schema:
            type: string
            enum: ["all", "0", "1", "2", "3"]
          default: "all"
          example: "all"
        - name: includeSources
          in: query
          required: false
          description: Include bibliographic source references (sourceId and page) for the tuning system
          schema:
            type: string
            enum: ["true", "false"]
          default: "true"
          example: "true"
        - name: includeArabic
          in: query
          required: false
          description: |
            Return bilingual responses with Arabic script when true.
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
          example: "true"
      responses:
        '200':
          description: Tuning system pitch classes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - tuningSystem
                  - startingNotes
                  - selectedStartingNote
                  - stats
                  - pitchClasses
                  - context
                  - links
                properties:
                  tuningSystem:
                    $ref: '#/components/schemas/EntityRef'
                  startingNotes:
                    type: object
                    properties:
                      idNames:
                        type: array
                        items:
                          type: string
                      displayNames:
                        type: array
                        items:
                          type: string
                  selectedStartingNote:
                    type: object
                    properties:
                      idName:
                        type: string
                      displayName:
                        type: string
                  stats:
                    type: object
                    properties:
                      totalPitchClasses:
                        type: integer
                      numberOfPitchClassesSingleOctave:
                        type: integer
                      referenceFrequency:
                        type: number
                      octaves:
                        type: array
                        items:
                          type: integer
                  pitchClasses:
                    type: array
                    description: Array of formatted pitch class data
                    items:
                      type: object
                  sources:
                    type: array
                    description: Bibliographic source references for this tuning system
                    items:
                      type: object
                      properties:
                        sourceId:
                          type: string
                          description: URL-safe source identifier
                          example: "al-Shawwa-(1946)"
                        page:
                          type: string
                          description: Page number or range where the tuning system is documented
                          example: "49"
                  context:
                    type: object
                    properties:
                      pitchClassDataType:
                        type: string
                      octave:
                        type: string
                        nullable: true
                        description: '"all" if all octaves, or specific octave number'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                      tuningSystem:
                        type: string
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tuning system not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /pitch-classes/note-names/{noteName}/compare:
    get:
      tags:
        - Pitch Classes
      summary: Compare pitch class by note name across tuning systems
      description: |
        Compare a pitch class (by note name) across different tuning systems.
        
        Returns comprehensive data for each tuning system including:
        - Pitch class data in the requested format
        - Context information (tuning system, starting note, reference frequency)
      operationId: comparePitchClassByNoteName
      parameters:
        - name: noteName
          in: path
          required: true
          description: URL-safe note name (e.g., "rast", "dugah", "nawa")
          schema:
            type: string
          example: "rast"
        - name: tuningSystems
          in: query
          required: true
          description: Comma-separated tuning system IDs
          schema:
            type: string
          example: "IbnSina-(1037),al-Farabi-(950g)"
        - name: startingNote
          in: query
          required: true
          description: |
            Starting note (applies to all tuning systems), or "all" to include all available starting notes for each tuning system
          schema:
            type: string
          example: "yegah"
        - name: pitchClassDataType
          in: query
          required: false
          description: Pitch class data format
          schema:
            type: string
            enum: [all, englishName, fraction, cents, decimalRatio, stringLength, frequency, abjadName, fretDivision, midiNoteNumber, midiNoteDeviation, centsDeviation, referenceNoteName]
            default: all
          example: "all"
        - name: includeArabic
          in: query
          required: false
          description: |
            Return bilingual responses with Arabic script when true.
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
          example: "true"
      responses:
        '200':
          description: Comparison data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: object
                    properties:
                      noteName:
                        $ref: '#/components/schemas/IdentifierRef'
                  comparisons:
                    type: array
                    description: Array of comparison results, one per tuning system/starting note combination
                    items:
                      type: object
                      properties:
                        tuningSystem:
                          $ref: '#/components/schemas/EntityRef'
                        startingNote:
                          $ref: '#/components/schemas/IdentifierRef'
                        pitchClass:
                          $ref: '#/components/schemas/PitchClass'
                        context:
                          type: object
                          properties:
                            referenceFrequency:
                              type: number
                            pitchClassDataType:
                              type: string
                  meta:
                    type: object
                    properties:
                      totalComparisons:
                        type: integer
                      successfulComparisons:
                        type: integer
                      failedComparisons:
                        type: integer
                  links:
                    $ref: '#/components/schemas/Links'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    PitchClass:
      type: object
      description: Pitch class data (fields vary by format parameter)
      properties:
        pitchClassIndex:
          type: integer
          description: Index in tuning system
          example: 0
        octave:
          type: integer
          description: Octave number (0, 1, 2, or 3) indicating which octave this pitch class belongs to
          example: 1
        scaleDegree:
          type: string
          description: Roman numeral maqām degree
          example: "I"
        noteName:
          type: string
          description: URL-safe note name
          example: "ushayran"
        noteNameDisplay:
          type: string
          description: Note name with diacritics
          example: "ʿushayrān"
        englishName:
          type: string
          description: English transliteration
          example: "ushayran"
        abjadName:
          type: string
          description: Arabic script name
          example: "عشيران"
        frequency:
          type: number
          format: float
          description: Frequency in Hz (when pitchClassDataType=frequency or all)
          example: 97.999
        cents:
          type: number
          format: float
          description: Cents above reference (when pitchClassDataType=cents or all)
          example: 0.0
        fraction:
          type: string
          description: Interval fraction ratio (when pitchClassDataType=fraction or all)
          example: "1/1"
        decimalRatio:
          type: number
          format: float
          description: Decimal frequency ratio (when pitchClassDataType=decimalRatio or all)
          example: 1.0
        stringLength:
          type: number
          format: float
          description: Relative string length (when pitchClassDataType=stringLength or all)
          example: 1.0
        midiNoteDecimal:
          type: number
          format: float
          description: MIDI note with decimal cents (when format=midi or all)
          example: 38.0
        midiNotePlusCentsDeviation:
          type: string
          description: MIDI note ± cents deviation (when format=all)
          example: "D2 +0.0"
        centsDeviation:
          type: number
          format: float
          description: Cents deviation from equal temperament (when format=all)
          example: 0.0
        ipnReferenceNoteName:
          type: string
          description: IPN reference note (when format=all)
          example: "D2"
    EntityRef:
      type: object
      required: [id, idName, displayName]
      properties:
        id:
          type: string
          description: Numeric or canonical identifier
        idName:
          type: string
          description: URL-safe identifier
        displayName:
          type: string
          description: Human-readable name with diacritics and transposition suffixes when applicable
    IdentifierRef:
      type: object
      required: [idName, displayName]
      properties:
        id:
          type: string
          description: Optional numeric or canonical identifier
        idName:
          type: string
          description: URL-safe identifier
          example: "rast"
        displayName:
          type: string
          description: Human-readable name with diacritics
          example: "rāst"
        displayNameAr:
          type: string
          description: Arabic display name (present when includeArabic=true)
          example: "راست"
    Links:
      type: object
      description: Navigation links for related endpoints
      properties:
        self:
          type: string
          description: Current endpoint URL
        detail:
          type: string
          description: Link to detail endpoint
        availability:
          type: string
          description: Link to availability endpoint
        compare:
          type: string
          description: Link to compare endpoint
        intervals:
          type: string
          description: Link to intervals endpoint
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to retrieve maqāmāt"
        message:
          type: string
          description: Additional detailed message explaining the error (not always present)
          example: "An unexpected error occurred while processing the request"
        details:
          type: string
          description: Additional error details (for some error types, not always present)
          example: "Database connection failed"
        hint:
          type: string
          description: Helpful hint for resolving the error (not always present)
          example: "Please try again later or contact support"
        availabilityUrl:
          type: string
          description: URL to check availability (only for missing parameter errors, not always present)
          example: "/api/maqamat/{id}/availability"
        validOptions:
          type: array
          description: List of valid values (only for invalid parameter errors, not always present). These are display values (with diacritics) as returned by the tuning system.
          items:
            type: string
          example: ["ʿushayrān", "yegāh"]
  responses:
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Failed to retrieve maqāmāt"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticated requests (if implemented in future)
security: []  # Currently no authentication required